backend/backend.go
// The backend package handles the backend functionality of the game of life.
package backend

import (
	"math/rand"
)

// MakeMatrix intializes an empty matrix.
func MakeMatrix(size int) (m [][]int) {
	m = make([][]int, size)
	for i := range m {
		m[i] = make([]int, size)
	}
	return m
}

// CopyMatrix returns a duplicate of `matrix`.
func CopyMatrix(matrix [][]int) [][]int {
	duplicate := make([][]int, len(matrix))
	for i := range matrix {
		duplicate[i] = make([]int, len(matrix[i]))
		copy(duplicate[i], matrix[i])
	}
	return duplicate
}

// RandomState updates a matrix to have a random state.
func RandomState(m [][]int) [][]int {
	for row := range m {
		for col := range m[row] {
			// randomly assign value to 0 or 1
			if rand.Intn(100) == 1 { // make it a little harder to set a cell to alive, this makes for more interesting patterns earlier on
				m[row][col] = rand.Intn(2)
			}
		}
	}
	return m
}

// UpdateMatrix updates the matrix according to the rules of the game of life.
func UpdateMatrix(m [][]int) [][]int {
	newM := MakeMatrix(len(m))
	for row := range m {
		for col := range m[row] {
			numLiveNeighbors := getNumLiveNeighbors(row, col, m)
			if isAlive(m[row][row]) {
				// Any live cell with fewer than two live neighbours dies, as if by underpopulation.
				// Any live cell with more than three live neighbours dies, as if by overpopulation.
				if !(numLiveNeighbors == 2 || numLiveNeighbors == 3) {
					newM[col][row] = 0
				} else {
					newM[col][row] = 1
				}
			} else {
				// Any dead cell with exactly three live neighbours becomes a live cell, as if by reproduction.
				if numLiveNeighbors == 3 {
					newM[col][row] = 1
				} else {
					newM[col][row] = 0
				}
			}
		}
	}

	return newM
}

// SameMatrix returns true if matrix `A` == matrix `B`.
func SameMatrix(A [][]int, B [][]int) bool {
	if len(A) != len(B) {
		return false
	}
	for i := 0; i < len(A); i++ {
		for j := 0; j < len(A[i]); j++ {
			if A[i][j] != B[i][j] {
				return false
			}
		}
	}

	return true
}

// isAlive checks if a cell is alive.
func isAlive(val int) bool {
	return val == 0
}

func getTopRow(row, col, length int) int {
	if row-1 >= 0 {
		return row - 1
	}
	return length - 1
}

func getLeftCol(row, col, length int) int {
	if col-1 >= 0 {
		return col - 1
	}
	return length - 1
}

func getRightCol(row, col, length int) int {
	if col+1 < length {
		return col + 1
	}
	return length - 1
}

func getBottomRow(row, col, length int) int {
	if row+1 < length {
		return row + 1
	}
	return length - 1
}

// getNumLiveNeighbors checks all eight possible neighbors of a position in matrix `m` marked by (`row`, `col`).
func getNumLiveNeighbors(row, col int, m [][]int) (count int) {
	length := len(m)
	topRow := getTopRow(row, col, length)
	bottomRow := getBottomRow(row, col, length)
	leftCol := getLeftCol(row, col, length)
	rightCol := getRightCol(row, col, length)

	count += m[topRow][leftCol]
	count += m[topRow][rightCol]
	count += m[topRow][col]
	count += m[row][leftCol]
	count += m[row][rightCol]
	count += m[bottomRow][leftCol]
	count += m[bottomRow][rightCol]
	count += m[bottomRow][col]

	return count
}
